requirements.txt
Django==3.0.6
mysqlclient==1.4.6
mutagen==1.44.0
django-compressor==2.4
django-libsass==0.8
django-solo==1.1.3